FWD/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1

# LINE Bot
line-bot-sdk==3.13.0

# AI Vision APIs (สามารถเลือกใช้อันใดอันหนึ่งหรือทั้งคู่)
google-genai>=0.3.0             # Gemini API - NEW package (google.generativeai is deprecated)
openai==1.57.0                  # OpenAI GPT-4o

# Data Validation
pydantic==2.10.3
pydantic[email]==2.10.3
pydantic-settings==2.6.0

# Image Processing
pillow==11.0.0

# HTTP Client
httpx==0.27.2
aiofiles==24.1.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0

# Deployment (optional)
gunicorn==23.0.0
